修复c拓展跳转指令错误

+++ src/core/rvinst.c
@@ -326,14 +326,14 @@ void inst_c_addi16sp(struct RiscvCore *core) {
         core->regs[2] = ((isize)core->regs[2] + imm);
 }
 void inst_c_j(struct RiscvCore *core) {
-    u32 offset = ((DEC.inst >> 2) & 0x1) << 5 | ((DEC.inst >> 3) & 0x7) << 1 | ((DEC.inst >> 6) & 0x1) << 7 |
+    usize offset = ((DEC.inst >> 2) & 0x1) << 5 | ((DEC.inst >> 3) & 0x7) << 1 | ((DEC.inst >> 6) & 0x1) << 7 |
                  ((DEC.inst >> 7) & 0x1) << 6 | ((DEC.inst >> 8) & 0x1) << 10 | ((DEC.inst >> 9) & 0x3) << 8 |
                  ((DEC.inst >> 11) & 0x1) << 4 | ((DEC.inst >> 12) & 0x1) << 11;
     offset |= (offset & 0x800) ? ~0x3ff : 0;
     DEC.next_pc = core->pc + offset;
 }
 void inst_c_jal(struct RiscvCore *core) {
-    u32 offset = ((DEC.inst >> 2) & 0x1) << 5 | ((DEC.inst >> 3) & 0x7) << 1 | ((DEC.inst >> 6) & 0x1) << 7 |
+    usize offset = ((DEC.inst >> 2) & 0x1) << 5 | ((DEC.inst >> 3) & 0x7) << 1 | ((DEC.inst >> 6) & 0x1) << 7 |
                  ((DEC.inst >> 7) & 0x1) << 6 | ((DEC.inst >> 8) & 0x1) << 10 | ((DEC.inst >> 9) & 0x3) << 8 |
                  ((DEC.inst >> 11) & 0x1) << 4 | ((DEC.inst >> 12) & 0x1) << 11;
     offset |= (offset & 0x800) ? ~0x3ff : 0;
@@ -341,14 +341,14 @@ void inst_c_jal(struct RiscvCore *core) {
     DEC.next_pc   = core->pc + offset;
 }
 void inst_c_beqz(struct RiscvCore *core) {
-    u32 offset = ((DEC.inst >> 2) & 0x1) << 5 | ((DEC.inst >> 3) & 0x3) << 1 | ((DEC.inst >> 5) & 0x3) << 6 |
+    usize offset = ((DEC.inst >> 2) & 0x1) << 5 | ((DEC.inst >> 3) & 0x3) << 1 | ((DEC.inst >> 5) & 0x3) << 6 |
                  ((DEC.inst >> 10) & 0x3) << 3 | ((DEC.inst >> 12) & 0x1) << 8;
     offset |= (offset & 0x100) ? ~0xff : 0;
     if (core->regs[DEC.rs1_] == 0)
         DEC.next_pc = core->pc + offset;
 }
 void inst_c_bnez(struct RiscvCore *core) {
-    u32 offset = ((DEC.inst >> 2) & 0x1) << 5 | ((DEC.inst >> 3) & 0x3) << 1 | ((DEC.inst >> 5) & 0x3) << 6 |
+    usize offset = ((DEC.inst >> 2) & 0x1) << 5 | ((DEC.inst >> 3) & 0x3) << 1 | ((DEC.inst >> 5) & 0x3) << 6 |
                  ((DEC.inst >> 10) & 0x3) << 3 | ((DEC.inst >> 12) & 0x1) << 8;
     offset |= (offset & 0x100) ? ~0xff : 0;
     if (core->regs[DEC.rs1_] != 0)
